correção de sintaxe e reorganização

+++ Projetos/spring_mte/src/main/java/com/mte2023/spring_mte/repositories/UserRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mte2023.spring_mte.entities.User;
 
-public interface UserRepository extends JpaRepository<User, Integer>{ }
+public interface UserRepository extends JpaRepository<User, Long>{ }
